[refactor] parecer: services front #2150
