+++ application/modules/parecer/Bootstrap.php
@@ -20,8 +20,8 @@ class Parecer_Bootstrap extends Zend_Application_Module_Bootstrap
                     'analise-inicial-consolidacao-rest',
                     'analise-inicial-outros-produtos-rest',
                     'analise-inicial-finalizacao-rest',
-                    'analise-inicial-historico-rest',
                     'analise-inicial-impedimento-rest',
+                    'historico-produto-rest',
                     'planilha-produto-rest',
                     'gerenciar-parecer-rest',
                     'gerenciar-distribuir-produto-rest',
