+++ application/modules/parecer/controllers/HistoricoProdutoRestController.php
@@ -1,6 +1,6 @@
 <?php
 
-class Parecer_AnaliseInicialHistoricoRestController extends MinC_Controller_Rest_Abstract
+class Parecer_HistoricoProdutoRestController extends MinC_Controller_Rest_Abstract
 {
 
     public function __construct(Zend_Controller_Request_Abstract $request, Zend_Controller_Response_Abstract $response, array $invokeArgs = array())
