+++ front/src/modules/parecer/service/AnaliseInicial.js
@@ -17,6 +17,8 @@ export const obterProdutoParaAnalise = (params) => {
     return api.getRequest(`/${MODULE}/${controller}${queryParams}`);
 };
 
+/** Análise de Conteúdo */
+
 export const obterAnaliseConteudo = (params) => {
     const controller = 'analise-inicial-conteudo-rest';
     const id = `id/${params.id}`;
@@ -26,10 +28,13 @@ export const obterAnaliseConteudo = (params) => {
 
     return api.getRequest(`/${MODULE}/${controller}${queryParams}`);
 };
+
 export const salvarAnaliseConteudo = (params) => {
     const controller = 'analise-inicial-conteudo-rest';
     return api.postRequest(`/${MODULE}/${controller}/`, api.buildData(params));
 };
+
+/** Análise de Custo */
 
 export const obterPlanilhaParaAnalise = (params) => {
     const controller = 'analise-inicial-custo-rest';
@@ -42,20 +47,12 @@ export const obterPlanilhaParaAnalise = (params) => {
     return api.getRequest(`/${MODULE}/${controller}${queryParams}`);
 };
 
-export const obterProdutosSecundarios = (params) => {
-    const controller = 'analise-inicial-outros-produtos-rest';
-    const idProduto = `id/${params.id}`;
-    const idPronac = `idPronac/${params.idPronac}`;
-
-    const queryParams = `/${idProduto}/${idPronac}`;
-
-    return api.getRequest(`/${MODULE}/${controller}${queryParams}`);
-};
-
 export const salvarAvaliacaoItem = (params) => {
     const controller = 'analise-inicial-custo-rest';
     return api.postRequest(`/${MODULE}/${controller}/`, api.buildData(params));
 };
+
+/** Consolidação */
 
 export const obterAnaliseConsolidacao = (params) => {
     const controller = 'analise-inicial-consolidacao-rest';
@@ -72,28 +69,16 @@ export const salvarAnaliseConsolidacao = (params) => {
     return api.postRequest(`/${MODULE}/${controller}`, api.buildData(params));
 };
 
+/** Finalização */
+
 export const finalizarAnalise = (params) => {
     const controller = 'analise-inicial-finalizacao-rest';
     return api.postRequest(`/${MODULE}/${controller}/`, api.buildData(params));
 };
 
-export const obterHistoricoProduto = (params) => {
-    const controller = 'analise-inicial-historico-rest';
-    const idProduto = `idProduto/${params.idProduto}`;
-    const idPronac = `idPronac/${params.idPronac}`;
-    const stPrincipal = `stPrincipal/${params.stPrincipal}`;
-
-    const queryParams = `/${idProduto}/${idPronac}/${stPrincipal}`;
-
-    return api.getRequest(`/${MODULE}/${controller}${queryParams}`);
-};
+/** Declarar Impedimento */
 
 export const salvarDeclaracaoImpedimento = (params) => {
     const controller = 'analise-inicial-impedimento-rest';
-    return api.postRequest(`/${MODULE}/${controller}`, api.buildData(params));
-};
-
-export const restaurarPlanilhaProduto = (params) => {
-    const controller = 'planilha-produto-rest';
     return api.postRequest(`/${MODULE}/${controller}`, api.buildData(params));
 };
