+++ front/src/modules/parecer/service/Planilha.js
@@ -1,0 +1,8 @@
+import * as api from '@/helpers/api/base';
+
+const MODULE = 'parecer';
+
+export const restaurarPlanilhaProduto = (params) => {
+    const controller = 'planilha-produto-rest';
+    return api.postRequest(`/${MODULE}/${controller}`, api.buildData(params));
+};
